refactor: keypairs api.
